Bugs :)

+++ Assets/MergeMonsterClone/Assets/Scripts/ScriptsO/ReFactor/WinScreenHandler.cs
@@ -25,7 +25,7 @@ public class WinScreenHandler : MonoBehaviour
     }
 
     //this func will added after bank system done
-    public void SetTexts(int level, int reward, int total)
+    public void SetTexts(int level, float reward, float total)
     {
         _level.text = level.ToString();
         _earnedGold.text = reward.ToString();
